+++ sdk/resourcemanager/consumption/armconsumption/tags_client.go
@@ -41,7 +41,7 @@ func NewTagsClient(credential azcore.TokenCredential, options *arm.ClientOptions
 // Get - Get all available tag keys for the defined scope
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2021-10-01
+// Generated from API version 2024-04-01
 //   - scope - The scope associated with tags operations. This includes '/subscriptions/{subscriptionId}/' for subscription scope,
 //     '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}' for
 //     resourceGroup scope, '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}' for Billing Account scope,
@@ -81,7 +81,7 @@ func (client *TagsClient) getCreateRequest(ctx context.Context, scope string, op
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-10-01")
+	reqQP.Set("api-version", "2024-04-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil